Fix: Top_k should be an integer

+++ openhands/core/config/llm_config.py
@@ -68,7 +68,7 @@ class LLMConfig(BaseModel):
     )  # maximum number of characters in an observation's content when sent to the llm
     temperature: float = Field(default=0.0)
     top_p: float = Field(default=1.0)
-    top_k: float | None = Field(default=None)
+    top_k: int | None = Field(default=None)
     custom_llm_provider: str | None = Field(default=None)
     max_input_tokens: int | None = Field(default=None)
     max_output_tokens: int | None = Field(default=None)
